+++ Changelog.md
@@ -1,3 +1,4 @@
 1. Baseline. Score: 43.75
 2. Taking all features that have less than 300 NAS in test set. Score: 38.4
 3. Taking all features that have less than 5000 NAS in test set. Score: 35.2
+4. With gridsearch and pipeline with standardscaler. Score: 35.55